requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
certifi==2024.12.14
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.115.6
fonttools==4.55.3
git-filter-repo==2.47.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httpcore==0.18.0
httpx==0.25.0
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.10.0
numpy==2.2.0
openaq==0.3.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pyparsing==3.2.0
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
rsa==4.9
seaborn==0.13.2
six==1.17.0
starlette==0.41.3
tenacity==9.0.0
tzdata==2024.2
uvicorn==0.34.0
